Implementando regras para farm

+++ FarmManagementSystem.Services/Services/FarmService.cs
@@ -1,11 +1,15 @@
 ﻿using FarmManagementSystem.Domain.Entities;
 using FarmManagementSystem.Domain.Interfaces;
+using FarmManagementSystem.Infra.Repositorios;
+using FarmManagementSystem.Services.Dtos;
+using System.ComponentModel.DataAnnotations;
 
 namespace FarmManagementSystem.Services.Services
 {
-    public class FarmService(IFarmRepository farmRepository)
+    public class FarmService(IFarmRepository farmRepository, ILocationRepository locationRepository)
     {
         private readonly IFarmRepository _farmRepository = farmRepository;
+        private readonly ILocationRepository _locationRepository = locationRepository;
 
         public List<Farm> GetAll()
         {
@@ -44,11 +48,13 @@ namespace FarmManagementSystem.Services.Services
             }
         }
 
-        public void Add(Farm user)
+        public void Add(FarmDto FarmDto)
         {
             try
             {
-                _farmRepository.Add(user);
+                var farm = ProcessingAddfields(FarmDto);
+
+                _farmRepository.Add(farm);
             }
             catch (Exception ex)
             {
@@ -56,18 +62,25 @@ namespace FarmManagementSystem.Services.Services
             }
         }
 
-        public void Update(Farm farm)
+        public void Update(FarmDto FarmDto)
         {
             try
             {
+                var farm = ProcessingUpdatefields(FarmDto);
                 var farmInDb = _farmRepository.GetById(farm.Id);
+
+                if (farmInDb.UserId != farm.UserId)
+                    throw new ValidationException("Usuário sem acesso para alterações nessa fazenda.");
+
+                if (!farmInDb.IsFarmActive())
+                    throw new ValidationException("Apenas fazendas ativas podem sofrar modificações.");
+
                 _farmRepository.Update(farmInDb, farm);
             }
             catch (Exception ex)
             {
                 throw new Exception(ex.Message);
             }
-
         }
 
         public void Delete(int id)
@@ -75,6 +88,12 @@ namespace FarmManagementSystem.Services.Services
             try
             {
                 var farmInDb = _farmRepository.GetById(id);
+                var farmLocationInDb = _locationRepository.GetById(farmInDb.Id);
+                //Processar a exclusão da localização...........................
+
+                if (farmInDb.IsFarmActive())
+                    throw new ValidationException("Apenas fazenda desativadas podem ser excluidas.");
+
                 _farmRepository.Delete(farmInDb);
             }
             catch (Exception ex)
@@ -82,5 +101,48 @@ namespace FarmManagementSystem.Services.Services
                 throw new Exception(ex.Message);
             }
         }
+
+        #region processingFields
+        private static Farm ProcessingAddfields(FarmDto farmDto)
+        {
+            var farm = new Farm
+            {
+                UserId = farmDto.UserId,
+                Name = farmDto.Name,
+                Description = farmDto.Description,
+                DateAdd = farmDto.DateAdd,
+                FarmIsActive = farmDto.FarmIsActive,
+                Location = new Location
+                {
+                    Latitude = farmDto.Location.Latitude,
+                    Longitude = farmDto.Location.Longitude,
+                }
+            };
+
+            farm.Validate();
+            farm.Location.Validate();
+
+            return farm;
+        }
+
+        private static Farm ProcessingUpdatefields(FarmDto farmDto)
+        {
+            var farm = new Farm
+            {
+                Id = farmDto.Id,
+                UserId = farmDto.UserId,
+                Name = farmDto.Name,
+                Description = farmDto.Description,
+                DateAdd = farmDto.DateAdd,
+                FarmIsActive = farmDto.FarmIsActive,
+            };
+
+            farm.ValidateId();
+            farm.Validate();
+
+            return farm;
+        }
+        #endregion
+
     }
 }
